backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
alembic
psycopg2-binary
pgvector
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.0.0
python-multipart
python-dotenv
boto3
pydantic-settings
fastembed
email-validator
zep-cloud
google-generativeai
groq
pytest
httpx
rembg[cpu]
qdrant-client==1.12.1
transformers
torch
pillow
pydantic[email]
requests
aiofiles
email-validator
azure-storage-blob
openai
langgraph
langchain
langchain-openai
langchain-community
langchain-groq
sentence-transformers
